fix(striker-ui): isolate file data converters, simplfy file list fetch
