+++ striker-ui/lib/api_converters/index.ts
@@ -3,6 +3,8 @@ import toAnvilMemoryCalcable from './toAnvilMemoryCalcable';
 import toAnvilOverviewHostList from './toAnvilOverviewHostList';
 import toAnvilOverviewList from './toAnvilOverviewList';
 import toAnvilSharedStorageOverview from './toAnvilSharedStorageOverview';
+import toFileDetail from './toFileDetail';
+import toFileOverviewList from './toFileOverview';
 
 export {
   toAnvilDetail,
@@ -10,4 +12,6 @@ export {
   toAnvilOverviewHostList,
   toAnvilOverviewList,
   toAnvilSharedStorageOverview,
+  toFileDetail,
+  toFileOverviewList,
 };
