+++ striker-ui/lib/api_converters/toFileOverview.ts
@@ -1,0 +1,16 @@
+const toFileOverviewList = (rows: string[][]) =>
+  rows.reduce<APIFileOverviewList>((previous, row) => {
+    const [uuid, name, size, type, checksum] = row;
+
+    previous[uuid] = {
+      checksum,
+      name,
+      size,
+      type: type as FileType,
+      uuid,
+    };
+
+    return previous;
+  }, {});
+
+export default toFileOverviewList;
